--- gesipan/src/main/java/com/zzangmin/gesipan/layer/basiccrud/entity/UserRole.java
@@ -1,5 +1,0 @@
-package com.zzangmin.gesipan.layer.basiccrud.entity;
-
-public enum UserRole {
-    NORMAL, ADMIN
-}
