+++ gesipan/src/main/java/com/zzangmin/gesipan/layer/login/entity/UserRole.java
@@ -1,0 +1,5 @@
+package com.zzangmin.gesipan.layer.login.entity;
+
+public enum UserRole {
+    NORMAL, ADMIN
+}
