+++ gesipan/src/main/java/com/zzangmin/gesipan/layer/login/interceptor/DevelopLoginCheckInterceptor.java
@@ -1,4 +1,4 @@
-package com.zzangmin.gesipan.interceptor;
+package com.zzangmin.gesipan.layer.login.interceptor;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Profile;
