+++ gesipan/src/main/java/com/zzangmin/gesipan/layer/login/interceptor/LoginCheckInterceptor.java
@@ -1,4 +1,4 @@
-package com.zzangmin.gesipan.interceptor;
+package com.zzangmin.gesipan.layer.login.interceptor;
 
 import org.springframework.web.servlet.HandlerInterceptor;
