+++ gesipan/src/main/java/com/zzangmin/gesipan/layer/login/service/GithubOauthService.java
@@ -1,9 +1,9 @@
-package com.zzangmin.gesipan.layer.basiccrud.service;
+package com.zzangmin.gesipan.layer.login.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.zzangmin.gesipan.layer.basiccrud.dto.oauth.GithubToken;
-import com.zzangmin.gesipan.layer.basiccrud.dto.oauth.UserResources;
-import com.zzangmin.gesipan.layer.basiccrud.entity.Users;
+import com.zzangmin.gesipan.layer.login.dto.oauth.GithubToken;
+import com.zzangmin.gesipan.layer.login.dto.oauth.UserResources;
+import com.zzangmin.gesipan.layer.login.entity.Users;
 
 public interface GithubOauthService {
