+++ gesipan/src/test/java/com/zzangmin/gesipan/web/service/DeployGithubOauthServiceTest.java
@@ -1,12 +1,12 @@
 package com.zzangmin.gesipan.web.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.zzangmin.gesipan.layer.basiccrud.repository.UsersRepository;
-import com.zzangmin.gesipan.layer.basiccrud.service.DevelopGithubOauthService;
-import com.zzangmin.gesipan.layer.basiccrud.dto.oauth.GithubToken;
-import com.zzangmin.gesipan.layer.basiccrud.dto.oauth.UserResources;
-import com.zzangmin.gesipan.layer.basiccrud.entity.UserRole;
-import com.zzangmin.gesipan.layer.basiccrud.entity.Users;
+import com.zzangmin.gesipan.layer.login.repository.UsersRepository;
+import com.zzangmin.gesipan.layer.login.service.DevelopGithubOauthService;
+import com.zzangmin.gesipan.layer.login.dto.oauth.GithubToken;
+import com.zzangmin.gesipan.layer.login.dto.oauth.UserResources;
+import com.zzangmin.gesipan.layer.login.entity.UserRole;
+import com.zzangmin.gesipan.layer.login.entity.Users;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
