+++ gesipan/src/test/java/com/zzangmin/gesipan/web/service/PostServiceTest.java
@@ -4,7 +4,9 @@ import com.zzangmin.gesipan.layer.basiccrud.entity.*;
 import com.zzangmin.gesipan.layer.basiccrud.repository.PostCategoryRepository;
 import com.zzangmin.gesipan.layer.basiccrud.repository.PostRepository;
 import com.zzangmin.gesipan.layer.basiccrud.repository.TemporaryPostRepository;
-import com.zzangmin.gesipan.layer.basiccrud.repository.UsersRepository;
+import com.zzangmin.gesipan.layer.login.entity.UserRole;
+import com.zzangmin.gesipan.layer.login.entity.Users;
+import com.zzangmin.gesipan.layer.login.repository.UsersRepository;
 import com.zzangmin.gesipan.layer.basiccrud.service.PostService;
 import com.zzangmin.gesipan.layer.basiccrud.dto.post.PostResponse;
 import com.zzangmin.gesipan.layer.basiccrud.dto.post.PostSaveRequest;
